refactor(sync): simplify transfer flow and UI state
Replace the old session/connect model with a simpler nearby-device transfer flow.
Devices are discovered as send targets only, while files and text are sent through
the raw TCP transfer path with clear offer, accept, decline, busy, and failure states.

Keep selected send items persistent after sending so users explicitly manage the
send tray. Treat pasted text as a first-class text transfer item instead of a
.txt file, and prevent outgoing clipboard history from being updated before the
receiver accepts the transfer.

Harden receiver-side transfer handling by blocking new incoming offers while the
receiver is requested, receiving, or received. Add receiver-busy handling, guard
quicksave so only known discovered peers can auto-send, and clean up failed raw
TCP setup by closing transfer state consistently.

Reshape presentation state into runtime, discovery, send, and receive slices so
the UI consumes predictable state instead of one flat god object. Split large
send, receive, and transfer composables into focused files and folders.

Update transfer failure messaging, runtime security copy, and raw TCP tests for
the new one-time-token transfer model.

Ignore local agent/tooling folders:
- .agents/skills/caveman/
- skills-main/
