+++ shared/src/commonMain/kotlin/com/liftley/sync360/features/sync/domain/model/FileTransferModels.kt
@@ -75,6 +75,7 @@ enum class TransferFailureReason {
     STORAGE_UNAVAILABLE,
     INTEGRITY_FAILED,
     RECEIVER_UNAVAILABLE,
+    RECEIVER_BUSY,
     NETWORK_FAILED,
     TIMED_OUT,
     WRITE_FAILED,
