+++ shared/src/commonMain/kotlin/com/liftley/sync360/features/sync/domain/model/SyncSnapshot.kt
@@ -2,8 +2,6 @@ package com.liftley.sync360.features.sync.domain.model
 
 data class SyncSnapshot(
     val runtime: SyncRuntimeState = SyncRuntimeState.Stopped,
-    val connection: ConnectionSnapshot = ConnectionSnapshot(),
-    val session: SessionSnapshot = SessionSnapshot.NoSession,
     val transfer: TransferSnapshot = TransferSnapshot(),
     val quickSaveEnabled: Boolean = false,
     val pendingIncomingOffer: PendingIncomingOffer? = null
