+++ shared/src/commonMain/kotlin/com/liftley/sync360/features/sync/presentation/ReceiveScreen.kt
@@ -12,6 +12,7 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.weight
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -59,6 +60,11 @@ import com.liftley.sync360.features.sync.presentation.components.FileTransferPro
 import com.liftley.sync360.features.sync.presentation.components.Sync360Surface
 import com.liftley.sync360.features.sync.presentation.components.SyncBottomNavigationBar
 import com.liftley.sync360.features.sync.presentation.components.formatBytes
+import com.liftley.sync360.features.sync.presentation.receive.ItemTransferStatus
+import com.liftley.sync360.features.sync.presentation.receive.ProgressFileRow
+import com.liftley.sync360.features.sync.presentation.receive.ReceiveProposalCard
+import com.liftley.sync360.features.sync.presentation.receive.ReceivedFileRow
+import com.liftley.sync360.features.sync.presentation.receive.getFileStatuses
 import com.liftley.sync360.features.sync.presentation.navigation.SyncRoute
 import kotlinx.coroutines.flow.Flow
 
@@ -87,7 +93,7 @@ fun ReceiveScreen(
         }
     }
 
-    val received = uiState.receivedFileBatch
+    val received = uiState.receive.receivedFileBatch
     val isSingleFile = received?.savedPaths?.size == 1
 
     Scaffold(
@@ -276,9 +282,9 @@ fun ReceiveScreen(
                 verticalArrangement = Arrangement.spacedBy(16.dp),
                 horizontalAlignment = Alignment.CenterHorizontally
             ) {
-                val offer = uiState.pendingIncomingOffer
-                val progress = uiState.fileTransferProgress
-                val failure = uiState.fileTransferFailure
+                val offer = uiState.receive.pendingIncomingOffer
+                val progress = uiState.receive.fileTransferProgress
+                val failure = uiState.receive.fileTransferFailure
 
                 when {
                     offer != null -> {
@@ -387,324 +393,6 @@ fun ReceiveScreen(
                     }
                 }
             }
-        }
-    }
-}
-
-@Composable
-private fun ReceiveProposalCard(
-    offer: PendingIncomingOffer,
-    onAccept: () -> Unit,
-    onDecline: () -> Unit
-) {
-    val colorScheme = MaterialTheme.colorScheme
-    val description = when (offer) {
-        is PendingIncomingOffer.Files ->
-            "${offer.fileCount} file${if (offer.fileCount == 1) "" else "s"} (${formatBytes(offer.totalBytes)})"
-        is PendingIncomingOffer.Text ->
-            "Text snippet: \"${offer.preview}\""
-    }
-
-    Sync360Surface(cornerRadius = 28.dp) {
-        Column(
-            modifier = Modifier.padding(24.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.spacedBy(16.dp)
-        ) {
-            Surface(
-                shape = CircleShape,
-                color = colorScheme.primaryContainer,
-                modifier = Modifier.size(64.dp)
-            ) {
-                Box(contentAlignment = Alignment.Center) {
-                    Icon(
-                        imageVector = if (offer is PendingIncomingOffer.Text) Icons.AutoMirrored.Filled.Article else Icons.Default.Inbox,
-                        contentDescription = null,
-                        tint = colorScheme.onPrimaryContainer,
-                        modifier = Modifier.size(28.dp)
-                    )
-                }
-            }
-
-            Column(horizontalAlignment = Alignment.CenterHorizontally, verticalArrangement = Arrangement.spacedBy(4.dp)) {
-                Text(
-                    text = "Incoming share from",
-                    style = MaterialTheme.typography.bodyLarge,
-                    color = colorScheme.onSurfaceVariant
-                )
-                Text(
-                    text = offer.senderName,
-                    style = MaterialTheme.typography.headlineSmall,
-                    fontWeight = FontWeight.Bold,
-                    color = colorScheme.onSurface
-                )
-            }
-
-            Surface(
-                shape = RoundedCornerShape(16.dp),
-                color = colorScheme.surfaceContainerHigh,
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                Column(
-                    modifier = Modifier.padding(16.dp),
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    Text(
-                        text = description,
-                        style = MaterialTheme.typography.bodyLarge,
-                        fontWeight = FontWeight.SemiBold,
-                        textAlign = TextAlign.Center
-                    )
-                }
-            }
-
-            Spacer(Modifier.height(8.dp))
-
-            Column(
-                modifier = Modifier.fillMaxWidth(),
-                verticalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                Button(
-                    onClick = onAccept,
-                    modifier = Modifier.fillMaxWidth(),
-                    shape = RoundedCornerShape(18.dp),
-                    contentPadding = PaddingValues(vertical = 14.dp)
-                ) {
-                    Text("Accept", fontWeight = FontWeight.Bold, style = MaterialTheme.typography.titleMedium)
-                }
-                OutlinedButton(
-                    onClick = onDecline,
-                    modifier = Modifier.fillMaxWidth(),
-                    shape = RoundedCornerShape(18.dp),
-                    contentPadding = PaddingValues(vertical = 14.dp),
-                    colors = ButtonDefaults.outlinedButtonColors(contentColor = colorScheme.error)
-                ) {
-                    Text("Decline", fontWeight = FontWeight.Bold, style = MaterialTheme.typography.titleMedium)
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun ReceivedFileRow(
-    file: TransferFilePreview,
-    path: String?,
-    onOpenFile: (String) -> Unit
-) {
-    val colorScheme = MaterialTheme.colorScheme
-    val isImage = file.mimeType.startsWith("image/")
-
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .then(if (path != null) Modifier.clickable { onOpenFile(path) } else Modifier)
-            .padding(horizontal = 16.dp, vertical = 12.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween
-    ) {
-        Row(
-            modifier = Modifier.weight(1f),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(12.dp)
-        ) {
-            Surface(
-                shape = RoundedCornerShape(8.dp),
-                color = colorScheme.primaryContainer,
-                modifier = Modifier.size(40.dp)
-            ) {
-                Box(contentAlignment = Alignment.Center) {
-                    if (isImage && path != null) {
-                        AsyncImage(
-                            model = path,
-                            contentDescription = null,
-                            modifier = Modifier.fillMaxSize(),
-                            contentScale = androidx.compose.ui.layout.ContentScale.Crop
-                        )
-                    } else {
-                        Icon(
-                            imageVector = if (file.mimeType.startsWith("image/") || file.mimeType.startsWith("video/")) {
-                                Icons.Default.PermMedia
-                            } else {
-                                Icons.Default.Description
-                            },
-                            contentDescription = null,
-                            tint = colorScheme.onPrimaryContainer,
-                            modifier = Modifier.size(22.dp)
-                        )
-                    }
-                }
-            }
-            Column(modifier = Modifier.weight(1f)) {
-                Text(
-                    text = file.name,
-                    style = MaterialTheme.typography.bodyMedium,
-                    fontWeight = FontWeight.Bold,
-                    color = colorScheme.onSurface,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis
-                )
-                Text(
-                    text = formatBytes(file.sizeBytes),
-                    style = MaterialTheme.typography.bodySmall,
-                    color = colorScheme.onSurfaceVariant
-                )
-            }
-        }
-
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(4.dp)
-        ) {
-            Text(
-                text = "Received",
-                style = MaterialTheme.typography.bodySmall,
-                fontWeight = FontWeight.Bold,
-                color = colorScheme.primary
-            )
-            Icon(
-                imageVector = Icons.Default.ChevronRight,
-                contentDescription = "Open file",
-                tint = colorScheme.onSurfaceVariant,
-                modifier = Modifier.size(18.dp)
-            )
-        }
-    }
-}
-
-@Composable
-private fun ProgressFileRow(
-    file: TransferFilePreview,
-    status: ItemTransferStatus
-) {
-    val colorScheme = MaterialTheme.colorScheme
-
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp, vertical = 12.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween
-    ) {
-        Row(
-            modifier = Modifier.weight(1f),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(12.dp)
-        ) {
-            Surface(
-                shape = RoundedCornerShape(8.dp),
-                color = when (status) {
-                    ItemTransferStatus.COMPLETED -> colorScheme.primaryContainer
-                    ItemTransferStatus.TRANSFERRING -> colorScheme.secondaryContainer
-                    ItemTransferStatus.QUEUED -> colorScheme.surfaceVariant
-                },
-                modifier = Modifier.size(40.dp)
-            ) {
-                Box(contentAlignment = Alignment.Center) {
-                    Icon(
-                        imageVector = if (file.mimeType.startsWith("image/") || file.mimeType.startsWith("video/")) {
-                            Icons.Default.PermMedia
-                        } else {
-                            Icons.Default.Description
-                        },
-                        contentDescription = null,
-                        tint = when (status) {
-                            ItemTransferStatus.COMPLETED -> colorScheme.onPrimaryContainer
-                            ItemTransferStatus.TRANSFERRING -> colorScheme.onSecondaryContainer
-                            ItemTransferStatus.QUEUED -> colorScheme.onSurfaceVariant
-                        },
-                        modifier = Modifier.size(22.dp)
-                    )
-                }
-            }
-            Column(modifier = Modifier.weight(1f)) {
-                Text(
-                    text = file.name,
-                    style = MaterialTheme.typography.bodyMedium,
-                    fontWeight = FontWeight.Bold,
-                    color = when (status) {
-                        ItemTransferStatus.QUEUED -> colorScheme.onSurfaceVariant.copy(alpha = 0.6f)
-                        else -> colorScheme.onSurface
-                    },
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis
-                )
-                Text(
-                    text = formatBytes(file.sizeBytes),
-                    style = MaterialTheme.typography.bodySmall,
-                    color = colorScheme.onSurfaceVariant
-                )
-            }
-        }
-
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(6.dp)
-        ) {
-            Text(
-                text = when (status) {
-                    ItemTransferStatus.COMPLETED -> "Received"
-                    ItemTransferStatus.TRANSFERRING -> "Transferring"
-                    ItemTransferStatus.QUEUED -> "Queued"
-                },
-                style = MaterialTheme.typography.bodySmall,
-                fontWeight = FontWeight.Bold,
-                color = when (status) {
-                    ItemTransferStatus.COMPLETED -> colorScheme.primary
-                    ItemTransferStatus.TRANSFERRING -> colorScheme.secondary
-                    ItemTransferStatus.QUEUED -> colorScheme.onSurfaceVariant.copy(alpha = 0.6f)
-                }
-            )
-
-            when (status) {
-                ItemTransferStatus.COMPLETED -> {
-                    Icon(
-                        imageVector = Icons.Default.Check,
-                        contentDescription = "Received",
-                        tint = colorScheme.primary,
-                        modifier = Modifier.size(16.dp)
-                    )
-                }
-                ItemTransferStatus.TRANSFERRING -> {
-                    CircularProgressIndicator(
-                        modifier = Modifier.size(14.dp),
-                        strokeWidth = 2.dp,
-                        color = colorScheme.secondary
-                    )
-                }
-                ItemTransferStatus.QUEUED -> {
-                    Icon(
-                        imageVector = Icons.Default.Schedule,
-                        contentDescription = "Queued",
-                        tint = colorScheme.onSurfaceVariant.copy(alpha = 0.6f),
-                        modifier = Modifier.size(16.dp)
-                    )
-                }
-            }
-        }
-    }
-}
-
-private enum class ItemTransferStatus {
-    QUEUED,
-    TRANSFERRING,
-    COMPLETED
-}
-
-private fun getFileStatuses(
-    files: List<TransferFilePreview>,
-    bytesTransferred: Long
-): List<ItemTransferStatus> {
-    var accumulatedBytes = 0L
-    return files.map { file ->
-        val fileStart = accumulatedBytes
-        val fileEnd = accumulatedBytes + file.sizeBytes
-        accumulatedBytes = fileEnd
-
-        when {
-            bytesTransferred >= fileEnd -> ItemTransferStatus.COMPLETED
-            bytesTransferred >= fileStart -> ItemTransferStatus.TRANSFERRING
-            else -> ItemTransferStatus.QUEUED
         }
     }
 }
