+++ shared/src/commonMain/kotlin/com/liftley/sync360/features/sync/presentation/SyncEvent.kt
@@ -4,22 +4,17 @@ import com.liftley.sync360.core.platform.FilePickerKind
 import com.liftley.sync360.features.sync.domain.model.PickedFile
 
 sealed interface SyncEvent {
-    data class AcceptConnection(val deviceId: String) : SyncEvent
-    data class DeclineConnection(val deviceId: String) : SyncEvent
     object ToggleQuickSave : SyncEvent
     data class AcceptIncomingOffer(val offerId: String) : SyncEvent
     data class DeclineIncomingOffer(val offerId: String) : SyncEvent
     data class CopyClipboard(val text: String) : SyncEvent
     data class UpdateOutgoingText(val text: String) : SyncEvent
-    data class RequestConnect(val deviceId: String) : SyncEvent
-    data class RequestConnectByHost(val hostAddress: String) : SyncEvent
-    object ConfirmConnect : SyncEvent
-    object DismissConnectRequest : SyncEvent
     object PasteFromClipboard : SyncEvent
     data class OpenFilePicker(val kind: FilePickerKind) : SyncEvent
     data class AddSelectedFiles(val files: List<PickedFile>) : SyncEvent
     data class AddCustomText(val text: String) : SyncEvent
-    object SendSelectedItems : SyncEvent
+
+    data class SendSelectedItemsToHost(val hostAddress: String) : SyncEvent
     data class SendSelectedItemsTo(val deviceId: String) : SyncEvent
     data class ProposeSendTo(val deviceId: String) : SyncEvent
     object CancelSendProposal : SyncEvent
