+++ shared/src/commonMain/kotlin/com/liftley/sync360/features/sync/presentation/SyncViewModel.kt
@@ -10,13 +10,9 @@ import com.liftley.sync360.features.sync.domain.model.SendItem
 import com.liftley.sync360.features.sync.domain.model.SyncProtocolLimits
 import com.liftley.sync360.features.sync.domain.repository.SyncRepository
 import com.liftley.sync360.features.sync.domain.runtime.SyncRuntimeController
-import com.liftley.sync360.features.sync.domain.controller.SyncConnectionController
 import com.liftley.sync360.features.sync.domain.controller.SyncTransferController
+
 import com.liftley.sync360.features.sync.domain.model.SyncRuntimeState
-import com.liftley.sync360.features.sync.domain.model.SessionSnapshot
-import com.liftley.sync360.features.sync.domain.model.SessionSecurityMode
-import com.liftley.sync360.features.sync.domain.model.ConnectionEvent
-import com.liftley.sync360.features.sync.domain.model.UserFacingFailure
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.launch
@@ -25,7 +21,6 @@ class SyncViewModel(
     val isDesktop: Boolean,
     private val repository: SyncRepository,
     private val runtimeController: SyncRuntimeController,
-    private val connectionController: SyncConnectionController,
     private val transferController: SyncTransferController,
     private val clipboardOperations: ClipboardOperations,
     private val fileOperations: FileOperations,
@@ -35,8 +30,10 @@ class SyncViewModel(
 
     private val _uiState = MutableStateFlow(
         SyncUiState(
-            localDeviceName = localDeviceName,
-            serverIp = localIpAddress
+            runtime = RuntimeUiState(
+                localDeviceName = localDeviceName,
+                serverIp = localIpAddress
+            )
         )
     )
     val uiState: StateFlow<SyncUiState> = _uiState.asStateFlow()
@@ -48,57 +45,37 @@ class SyncViewModel(
             runtimeController.snapshot.collect { snapshot ->
                 _uiState.update {
                     it.copy(
-                        runtimeState = snapshot.runtime,
-                        securityMode = (snapshot.session as? SessionSnapshot.Approved)
-                            ?.securityMode
-                            ?: SessionSecurityMode.TRUSTED_LAN_PLAINTEXT,
-                        pendingIncomingOffer = snapshot.pendingIncomingOffer,
-                        quickSaveEnabled = snapshot.quickSaveEnabled,
-                        fileTransferProgress = snapshot.transfer.progress,
-                        fileTransferFailure = snapshot.transfer.failure,
-                        receivedFileBatch = snapshot.transfer.receivedBatch,
-                        localNetworkHealthy = snapshot.runtime is SyncRuntimeState.Ready
+                        runtime = it.runtime.copy(
+                            runtimeState = snapshot.runtime,
+                            localNetworkHealthy = snapshot.runtime is SyncRuntimeState.Ready
+                        ),
+                        receive = it.receive.copy(
+                            pendingIncomingOffer = snapshot.pendingIncomingOffer,
+                            quickSaveEnabled = snapshot.quickSaveEnabled,
+                            fileTransferProgress = snapshot.transfer.progress,
+                            fileTransferFailure = snapshot.transfer.failure,
+                            receivedFileBatch = snapshot.transfer.receivedBatch
+                        )
                     )
                 }
             }
         }
         viewModelScope.launch {
             repository.nearbyDevices.collect { nearby ->
-                _uiState.update { it.copy(nearbyDevices = nearby) }
+                _uiState.update { it.copy(discovery = it.discovery.copy(nearbyDevices = nearby)) }
+            }
+        }
+        viewModelScope.launch {
+            repository.clipboardHistory.collect { history ->
+                _uiState.update { it.copy(send = it.send.copy(latestTexts = history)) }
             }
         }
         viewModelScope.launch {
             repository.isScanning.collect { scanning ->
-                _uiState.update { it.copy(isScanningForDevices = scanning) }
+                _uiState.update { it.copy(discovery = it.discovery.copy(isScanningForDevices = scanning)) }
             }
         }
-        viewModelScope.launch {
-            repository.connectionEvents.collect { event ->
-                when (event) {
-                    is ConnectionEvent.WaitingForApproval ->
-                        showMessage("Waiting for ${event.deviceName}")
-                    is ConnectionEvent.Connected ->
-                        showMessage("${event.deviceName} is available")
-                    is ConnectionEvent.Failed ->
-                        showMessage(event.toUiMessage())
-                }
-            }
-        }
-        viewModelScope.launch {
-            repository.sessionMessages.collect { messages ->
-                val texts = messages.filter { !it.isFile && !it.isFromMe }
-                    .takeLast(3) // LATEST_TEXT_LIMIT = 3
-                    .asReversed()
-                    .map { message ->
-                        ClipboardEntry(
-                            text = message.text,
-                            updatedLabel = message.timestamp.toHourMinuteLabel(),
-                            isFromMe = message.isFromMe
-                        )
-                    }
-                _uiState.update { it.copy(latestTexts = texts) }
-            }
-        }
+
     }
 
     private fun Long.toHourMinuteLabel(): String =
@@ -107,17 +84,9 @@ class SyncViewModel(
     fun onEvent(event: SyncEvent) {
         when (event) {
             is SyncEvent.UpdateOutgoingText -> {
-                _uiState.update { it.copy(outgoingText = event.text) }
+                _uiState.update { it.copy(send = it.send.copy(outgoingText = event.text)) }
             }
-            is SyncEvent.RequestConnect -> {
-                val device = findDevice(event.deviceId) ?: return
-                connectionController.request(device)
-            }
-            is SyncEvent.RequestConnectByHost -> connectionController.requestByHost(event.hostAddress)
-            is SyncEvent.ConfirmConnect -> connectionController.confirmRequest()
-            is SyncEvent.DismissConnectRequest -> connectionController.dismissRequest()
-            is SyncEvent.AcceptConnection -> connectionController.acceptIncoming(event.deviceId)
-            is SyncEvent.DeclineConnection -> connectionController.declineIncoming(event.deviceId)
+
             is SyncEvent.CopyClipboard -> {
                 if (event.text.isNotBlank()) {
                     clipboardOperations.writeClipboard(event.text)
@@ -127,7 +96,7 @@ class SyncViewModel(
             is SyncEvent.PasteFromClipboard -> {
                 val clipText = clipboardOperations.readClipboard()
                 if (!clipText.isNullOrBlank()) {
-                    _uiState.update { it.copy(outgoingText = clipText) }
+                    _uiState.update { it.copy(send = it.send.copy(outgoingText = clipText)) }
                 }
             }
             is SyncEvent.OpenFilePicker -> {
@@ -136,54 +105,59 @@ class SyncViewModel(
                 }
             }
             is SyncEvent.AddSelectedFiles -> {
-                val merged = (_uiState.value.selectedItems + event.files.map { SendItem.File(it) })
+                val merged = (_uiState.value.send.selectedItems + event.files.map { SendItem.File(it) })
                     .take(SyncProtocolLimits.MAX_FILES_PER_TRANSFER)
-                _uiState.update { it.copy(selectedItems = merged) }
+                _uiState.update { it.copy(send = it.send.copy(selectedItems = merged)) }
             }
             is SyncEvent.AddCustomText -> {
                 if (event.text.isNotBlank()) {
+                    if (event.text.length > SyncProtocolLimits.MAX_TEXT_LENGTH) {
+                        showMessage("Text is too large")
+                        return
+                    }
                     val textItem = event.text.toSendTextItem()
-                    val merged = (_uiState.value.selectedItems + textItem)
+                    val merged = (_uiState.value.send.selectedItems + textItem)
                         .take(SyncProtocolLimits.MAX_FILES_PER_TRANSFER)
-                    _uiState.update { it.copy(selectedItems = merged, outgoingText = "") }
+                    _uiState.update { it.copy(send = it.send.copy(selectedItems = merged, outgoingText = "")) }
                 }
             }
-            SyncEvent.SendSelectedItems -> {
-                val selected = _uiState.value.selectedItems
-                if (selected.isNotEmpty()) {
-                    transferController.sendItems(selected)
-                }
-                _uiState.update { it.copy(selectedItems = emptyList()) }
-            }
+
             is SyncEvent.ProposeSendTo -> {
-                if (_uiState.value.selectedItems.isEmpty()) {
+                if (_uiState.value.send.selectedItems.isEmpty()) {
                     showMessage("Add files or text first")
                     return
                 }
                 val target = findDevice(event.deviceId)
                 if (target != null) {
-                    _uiState.update { it.copy(pendingOutgoingOfferTarget = target) }
+                    _uiState.update { it.copy(send = it.send.copy(pendingOutgoingOfferTarget = target)) }
                 }
             }
             SyncEvent.CancelSendProposal -> {
-                _uiState.update { it.copy(pendingOutgoingOfferTarget = null) }
+                _uiState.update { it.copy(send = it.send.copy(pendingOutgoingOfferTarget = null)) }
             }
             is SyncEvent.SendSelectedItemsTo -> {
-                _uiState.update { it.copy(pendingOutgoingOfferTarget = null) }
-                val selected = _uiState.value.selectedItems
+                _uiState.update { it.copy(send = it.send.copy(pendingOutgoingOfferTarget = null)) }
+                val selected = _uiState.value.send.selectedItems
                 if (selected.isEmpty()) {
                     showMessage("Add files or text first")
                     return
                 }
                 transferController.sendItemsTo(event.deviceId, selected)
-                _uiState.update { it.copy(selectedItems = emptyList()) }
+            }
+            is SyncEvent.SendSelectedItemsToHost -> {
+                val selected = _uiState.value.send.selectedItems
+                if (selected.isEmpty()) {
+                    showMessage("Add files or text first")
+                    return
+                }
+                transferController.sendItemsToHost(event.hostAddress, selected)
             }
             SyncEvent.ClearSelectedItems -> {
-                _uiState.update { it.copy(selectedItems = emptyList()) }
+                _uiState.update { it.copy(send = it.send.copy(selectedItems = emptyList())) }
             }
             is SyncEvent.RemoveSelectedItem -> {
-                val updated = _uiState.value.selectedItems.filter { it.id != event.itemId }
-                _uiState.update { it.copy(selectedItems = updated) }
+                val updated = _uiState.value.send.selectedItems.filter { it.id != event.itemId }
+                _uiState.update { it.copy(send = it.send.copy(selectedItems = updated)) }
             }
             SyncEvent.DismissReceivedFiles -> transferController.dismissReceived()
             SyncEvent.DismissTransferFailure -> transferController.dismissFailure()
@@ -214,12 +188,12 @@ class SyncViewModel(
             }
             is SyncEvent.AcceptIncomingOffer -> repository.acceptIncomingOffer(event.offerId)
             is SyncEvent.DeclineIncomingOffer -> repository.declineIncomingOffer(event.offerId)
-            SyncEvent.ToggleQuickSave -> repository.setQuickSaveEnabled(!_uiState.value.quickSaveEnabled)
+            SyncEvent.ToggleQuickSave -> repository.setQuickSaveEnabled(!_uiState.value.receive.quickSaveEnabled)
         }
     }
 
     private fun findDevice(deviceId: String): DeviceProfile? =
-        _uiState.value.nearbyDevices.firstOrNull { it.id == deviceId }
+        _uiState.value.discovery.nearbyDevices.firstOrNull { it.id == deviceId }
 
     private fun showMessage(message: String) {
         _uiEffects.trySend(SyncUiEffect.ShowMessage(message))
@@ -237,41 +211,5 @@ class SyncViewModel(
             text = this,
             preview = shortPreview
         )
-    }
-}
-
-private fun ConnectionEvent.Failed.toUiMessage(): String {
-    val target = peer ?: "device"
-    return when (reason) {
-        UserFacingFailure.SERVER_UNAVAILABLE ->
-            "Sync360 could not open or reach the sharing server"
-        UserFacingFailure.INVALID_HOST ->
-            "Enter a valid local IPv4 address or hostname, optionally with a port"
-        UserFacingFailure.MISSING_ROUTE ->
-            "This device does not have a reachable address"
-        UserFacingFailure.UNREACHABLE ->
-            "Could not reach $target. Check Wi-Fi and firewall settings."
-        UserFacingFailure.REQUEST_TIMEOUT ->
-            "$target did not respond"
-        UserFacingFailure.APPROVAL_TIMEOUT ->
-            "$target is not ready"
-        UserFacingFailure.DECLINED ->
-            "Request declined"
-        UserFacingFailure.PEER_BUSY ->
-            "$target is busy"
-        UserFacingFailure.PROTOCOL_MISMATCH ->
-            "$target uses an incompatible Sync360 version"
-        UserFacingFailure.CLIENT_CLOSED ->
-            "Sync360 networking is stopped"
-        UserFacingFailure.TEXT_INVALID ->
-            "Text is empty or too large"
-        UserFacingFailure.TEXT_DELIVERY_FAILED ->
-            "Text could not be delivered"
-        UserFacingFailure.INCOMING_REQUEST_EXPIRED ->
-            "Incoming request expired"
-        UserFacingFailure.RECEIVER_DECLINED ->
-            "Receiver declined the transfer."
-        UserFacingFailure.UNKNOWN ->
-            "Device is not ready. Try again."
     }
 }
