+++ shared/src/commonMain/kotlin/com/liftley/sync360/features/sync/presentation/components/transfer/ReceivedFileBatchCard.kt
@@ -1,0 +1,87 @@
+package com.liftley.sync360.features.sync.presentation.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Close
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import com.liftley.sync360.features.sync.domain.model.ReceivedFileBatch
+import com.liftley.sync360.features.sync.presentation.SyncEvent
+@Composable
+fun ReceivedFileBatchCard(
+    batch: ReceivedFileBatch,
+    onEvent: (SyncEvent) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    val colorScheme = MaterialTheme.colorScheme
+    Sync360Surface(modifier = modifier, cornerRadius = 24.dp) {
+        Column(Modifier.padding(18.dp), verticalArrangement = Arrangement.spacedBy(14.dp)) {
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Text(
+                    text = batch.senderName,
+                    style = MaterialTheme.typography.titleMedium,
+                    fontWeight = FontWeight.Bold,
+                    color = colorScheme.onSurface
+                )
+                IconButton(
+                    onClick = { onEvent(SyncEvent.DismissReceivedFiles) },
+                    modifier = Modifier.size(24.dp)
+                ) {
+                    Icon(
+                        imageVector = Icons.Default.Close,
+                        contentDescription = "Dismiss",
+                        tint = colorScheme.onSurfaceVariant
+                    )
+                }
+            }
+            Text(
+                text = "Saved to Downloads",
+                style = MaterialTheme.typography.bodyMedium,
+                color = colorScheme.onSurfaceVariant
+            )
+            TransferPreviewSummaryRow(
+                title = "${batch.files.size} file${if (batch.files.size == 1) "" else "s"} saved",
+                subtitle = formatBytes(batch.files.sumOf { it.sizeBytes }),
+                files = batch.files
+            )
+            Button(
+                onClick = {
+                    if (batch.savedPaths.size == 1) {
+                        onEvent(SyncEvent.OpenFile(batch.savedPaths.first()))
+                    } else {
+                        onEvent(SyncEvent.OpenDownloadsFolder)
+                    }
+                },
+                modifier = Modifier.fillMaxWidth(),
+                shape = RoundedCornerShape(18.dp),
+                colors = ButtonDefaults.buttonColors(
+                    containerColor = colorScheme.primary,
+                    contentColor = colorScheme.onPrimary
+                ),
+                contentPadding = PaddingValues(vertical = 12.dp)
+            ) {
+                Text(if (batch.files.size == 1) "Open File" else "Show in folder", fontWeight = FontWeight.Bold)
+            }
+        }
+    }
+}
