+++ shared/src/commonMain/kotlin/com/liftley/sync360/features/sync/presentation/state/DiscoveryUiState.kt
@@ -1,0 +1,8 @@
+package com.liftley.sync360.features.sync.presentation
+
+import com.liftley.sync360.features.sync.domain.model.DeviceProfile
+
+data class DiscoveryUiState(
+    val nearbyDevices: List<DeviceProfile> = emptyList(),
+    val isScanningForDevices: Boolean = false
+)
